Added a WIP function to fetch and format webmentions
it (hopefully) won't work for now, until the WEBMENTION_TOKEN env var is set
It also does not actually push changes or delete the respository just yet because I'm still testing.
I need to also adjust this so that likes, etc, get added to the original object, rather than creating new ones. (mentions are okay to create new ones, as are replies — though replies should probably also be added to the original content...)